Added wrapper for AppUserTurnstile event. Move classes to correct packages. Fix docs nits.
