+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/DepartEventHandler.kt
@@ -2,11 +2,11 @@ package com.mapbox.services.android.navigation.v5.internal.navigation
 
 import android.content.Context
 import com.mapbox.services.android.navigation.v5.internal.location.MetricsLocation
+import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.MetricsReporter
 import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.NavigationEventFactory
 import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.PhoneState
 import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.SessionState
 import com.mapbox.services.android.navigation.v5.internal.navigation.routeprogress.MetricsRouteProgress
-import com.mapbox.services.android.navigation.v5.navigation.metrics.MetricsReporter
 
 internal class DepartEventHandler(
     private val applicationContext: Context,
