+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/MetricEvent.kt
@@ -1,4 +1,4 @@
-package com.mapbox.navigation.base.metrics
+package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
 
 import androidx.annotation.StringDef
 import com.google.gson.Gson
