+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/MetricsReporter.kt
@@ -1,8 +1,9 @@
-package com.mapbox.navigation.base.metrics
+package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
 
 /**
  * Interface for handling metric events.
- * @since 1.0.0
+ *
+ * @since 0.43.0
  */
 interface MetricsReporter {
 
@@ -10,7 +11,7 @@ interface MetricsReporter {
      * Add event to metrics reporter when this event occurs.
      *
      * @param metricEvent event that should be handled
-     * @since 1.0.0
+     * @since 0.43.0
      */
     fun addEvent(metricEvent: MetricEvent)
 
@@ -18,7 +19,7 @@ interface MetricsReporter {
      * Add observer that triggered when metric event handled
      *
      * @param metricsObserver metric event handle observer
-     * @since 1.0.0
+     * @since 0.43.0
      */
     fun setMetricsObserver(metricsObserver: MetricsObserver)
 }
