+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/NavigationAppUserTurnstileEvent.kt
@@ -1,0 +1,14 @@
+package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
+
+import com.google.gson.Gson
+import com.mapbox.android.telemetry.AppUserTurnstile
+
+internal class NavigationAppUserTurnstileEvent(
+    val event: AppUserTurnstile
+) : MetricEvent {
+
+    override val metric: String
+        get() = NavigationMetrics.APP_USER_TURNSTILE
+
+    override fun toJson(gson: Gson): String = gson.toJson(this)
+}
