+++ libandroid-navigation/src/test/java/com/mapbox/services/android/navigation/v5/internal/navigation/InitialGpsEventFactoryTest.java
@@ -1,6 +1,6 @@
 package com.mapbox.services.android.navigation.v5.internal.navigation;
 
-import com.mapbox.services.android.navigation.v5.navigation.metrics.MetricsReporter;
+import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.MetricsReporter;
 
 import org.junit.Test;
