+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/internal/metrics/MetricEvent.kt
@@ -1,4 +1,4 @@
-package com.mapbox.services.android.navigation.v5.navigation.metrics
+package com.mapbox.navigation.base.internal.metrics
 
 import androidx.annotation.StringDef
 import com.google.gson.Gson
