+++ libnavigation-metrics/src/main/java/com/mapbox/navigation/metrics/MapboxMetricsReporter.kt
@@ -2,17 +2,16 @@ package com.mapbox.navigation.metrics
 
 import android.content.Context
 import com.google.gson.Gson
-import com.mapbox.android.telemetry.AppUserTurnstile
 import com.mapbox.android.telemetry.MapboxTelemetry
-import com.mapbox.navigation.base.metrics.MetricEvent
-import com.mapbox.navigation.base.metrics.MetricsObserver
-import com.mapbox.navigation.base.metrics.MetricsReporter
-import com.mapbox.navigation.base.metrics.NavigationMetrics
-import com.mapbox.navigation.metrics.extensions.toTelemetryEvent
+import com.mapbox.navigation.base.internal.metrics.MetricEvent
+import com.mapbox.navigation.base.internal.metrics.MetricsObserver
+import com.mapbox.navigation.base.internal.metrics.MetricsReporter
+import com.mapbox.navigation.metrics.internal.utils.extensions.toTelemetryEvent
 import com.mapbox.navigation.utils.thread.WorkThreadHandler
 
 /**
  * Default implementation of [MetricsReporter] interface.
+ *
  * @since 1.0.0
  */
 object MapboxMetricsReporter : MetricsReporter {
@@ -62,14 +61,6 @@ object MapboxMetricsReporter : MetricsReporter {
     fun disable() {
         mapboxTelemetry.disable()
         threadWorker.stop()
-    }
-
-    fun addAppTurnstileEvent(event: AppUserTurnstile) {
-        mapboxTelemetry.push(event)
-
-        threadWorker.post {
-            metricsObserver?.onMetricUpdated(NavigationMetrics.APP_USER_TURNSTILE, gson.toJson(event))
-        }
     }
 
     override fun addEvent(metricEvent: MetricEvent) {
