+++ libnavigation-metrics/src/main/java/com/mapbox/navigation/metrics/internal/utils/extensions/MetricEx.kt
@@ -1,10 +1,10 @@
-package com.mapbox.services.android.navigation.v5.internal.utils.extensions
+package com.mapbox.navigation.metrics.internal.utils.extensions
 
 import com.mapbox.android.telemetry.Event
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.DirectionsMetrics
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.MetricEvent
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.NavigationAppUserTurnstileEvent
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.NavigationMetrics
+import com.mapbox.navigation.base.internal.metrics.DirectionsMetrics
+import com.mapbox.navigation.base.internal.metrics.MetricEvent
+import com.mapbox.navigation.base.internal.metrics.NavigationMetrics
+import com.mapbox.navigation.metrics.internal.NavigationAppUserTurnstileEvent
 
 internal fun MetricEvent.toTelemetryEvent(): Event? =
     when (metric) {
